Recreate via_stream as the new name for typed_via_stream
PR #538 deleted `unifex::via_stream()`, which was an untyped _Stream_ algorithm,
and changed all uses to use `unifex::typed_via_stream()` instead. This diff
recreates the `via_stream.hpp` header; it now defines the "typed via stream"
algorithm under the name `unifex::via_stream()`. At the same time, this diff
changes `unifex::typed_via_stream` to be a deprecated alias for the `via_stream`
algorithm.
