+++ examples/for_each_via_thread_scheduler.cpp
@@ -19,7 +19,7 @@
 #include <unifex/sync_wait.hpp>
 #include <unifex/then.hpp>
 #include <unifex/transform_stream.hpp>
-#include <unifex/typed_via_stream.hpp>
+#include <unifex/via_stream.hpp>
 
 #include <cstdio>
 
@@ -30,7 +30,7 @@ int main() {
 
   sync_wait(then(
       for_each(
-          typed_via_stream(
+          via_stream(
               context.get_scheduler(),
               transform_stream(
                   range_stream{0, 10},
