+++ examples/get_scheduler.cpp
@@ -20,7 +20,7 @@
 #include <unifex/sync_wait.hpp>
 #include <unifex/timed_single_thread_context.hpp>
 #include <unifex/transform_stream.hpp>
-#include <unifex/typed_via_stream.hpp>
+#include <unifex/via_stream.hpp>
 #include <unifex/with_query_value.hpp>
 
 #include <chrono>
@@ -45,7 +45,7 @@ int main() {
   // composed operations.
   sync_wait(with_query_value(
       then(
-          for_each(typed_via_stream(current_scheduler,
+          for_each(via_stream(current_scheduler,
                               transform_stream(range_stream{0, 10},
                                                [](int value) {
                                                  return value * value;
