+++ include/unifex/typed_via_stream.hpp
@@ -15,46 +15,12 @@
  */
 #pragma once
 
-#include <unifex/adapt_stream.hpp>
-#include <unifex/scheduler_concepts.hpp>
-#include <unifex/via.hpp>
-
-#include <unifex/detail/prologue.hpp>
+#include <unifex/via_stream.hpp>
 
 namespace unifex {
-namespace _typed_via_stream {
-  struct _fn {
-    template (typename StreamSender, typename Scheduler)
-        (requires scheduler<Scheduler>)
-    auto operator()(Scheduler&& scheduler, StreamSender&& stream) const {
-      return adapt_stream(
-          (StreamSender &&) stream,
-          [s = (Scheduler &&) scheduler](auto&& sender) mutable {
-            return via((decltype(sender))sender, s);
-          });
-    }
-    template (typename StreamSender, typename Scheduler)
-        (requires scheduler<Scheduler>)
-    auto operator()(StreamSender&& stream, Scheduler&& scheduler) const {
-      return adapt_stream(
-          (StreamSender &&) stream,
-          [s = (Scheduler &&) scheduler](auto&& sender) mutable {
-            return via((decltype(sender))sender, s);
-          });
-    }
-    template(typename Scheduler)
-        (requires scheduler<Scheduler>)
-    constexpr auto operator()(Scheduler&& scheduler) const
-        noexcept(is_nothrow_callable_v<
-          tag_t<bind_back>, _fn, Scheduler>)
-        -> bind_back_result_t<_fn, Scheduler> {
-      return bind_back(*this, (Scheduler&&)scheduler);
-    }
-  };
-} // namespace _typed_via_stream
 
-inline constexpr _typed_via_stream::_fn typed_via_stream {};
+[[deprecated(
+    "Use unifex::via_stream instead")]] inline constexpr _via_stream::_fn
+    typed_via_stream{};
 
-} // namespace unifex
-
-#include <unifex/detail/epilogue.hpp>
+}  // namespace unifex
