+++ test/on_stream_test.cpp
@@ -21,7 +21,7 @@
 #include <unifex/sync_wait.hpp>
 #include <unifex/then.hpp>
 #include <unifex/transform_stream.hpp>
-#include <unifex/typed_via_stream.hpp>
+#include <unifex/via_stream.hpp>
 
 #include <gtest/gtest.h>
 
@@ -35,7 +35,7 @@ TEST(on_stream, Smoke) {
 
   sync_wait(then(
       for_each(
-          typed_via_stream(
+          via_stream(
               context1.get_scheduler(),
               on_stream(
                   context2.get_scheduler(),
@@ -54,7 +54,7 @@ TEST(on_stream, Pipeable) {
     | transform_stream(
         [](int value) { return value * value; })
     | on_stream(context2.get_scheduler())
-    | typed_via_stream(context1.get_scheduler())
+    | via_stream(context1.get_scheduler())
     | for_each([](int value) { std::printf("got %i\n", value); })
     | then([]() { std::printf("done\n"); })
     | sync_wait();
